pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_utils"
version = "0.1.0"
description = "A unified interface for managing LLMs like Google Generative AI, Hugging Face, and OpenAI."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "OzLevi", email = "ozmaatuk@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "google-generativeai>=0.8.3",
    "openai>=1.57.4",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.6",
    "pydantic>=2.10.3"
]

[project.urls]
homepage = "https://github.com/OzMaatuk/LLMUtils"